refactor date picker into seperate component
